setup github action to attempt prebuild on linux machine
The prebuild script doesn't work on my machine because my XCode version doesn't support libstdc++. I've spent a bit of time trying to work around this, but to no luck. It seems like it's down to the fact that the 'nan' module includes a specific 'algorithm' import, and that import requires the old standard lib which is only available from XCode version 9 or lower. Github actions only goes as low as v10. So I'm instead following the linux instructions, hoping we can still get a prebuild done that way.